Moved stumbling to mozstumbler folder. So in gonk it will be gonk/mozstumbler and all the files will go in there
