+++ mozstumbler/MozStumbler.h
@@ -44,42 +44,5 @@ private:
 NS_IMPL_ISUPPORTS(StumblerInfo, nsICellInfoListCallback, nsIWifiScanResultsReady)
 
 
-class UploadStumbleRunnable final : public nsRunnable
-{
-public:
-  UploadStumbleRunnable()
-  {}
-
-  void TryToUploadFile();
-
-  NS_IMETHOD Run()
-  {
-    MOZ_ASSERT(NS_IsMainThread());
-    nsContentUtils::LogMessageToConsole("UploadStumbleRunnable\n");
-    TryToUploadFile();
-    return NS_OK;
-  }
-
-private:
-  virtual ~UploadStumbleRunnable() {}
-};
-
-
-class UploadEventListener : public nsIDOMEventListener
-{
-public:
-  explicit UploadEventListener(int64_t aFileSize)
-  : mFileSize(aFileSize)
-  {}
-
-/*interfaces for addref and release and queryinterface*/
-  NS_DECL_ISUPPORTS
-  NS_DECL_NSIDOMEVENTLISTENER
-
- protected:
-  virtual ~UploadEventListener() {}
-  int64_t mFileSize;
-};
-
 #endif // mozilla_system_mozstumbler_h__
