+++ mozstumbler/StumblerLogging.cpp
@@ -1,0 +1,7 @@
+#include "StumblerLogging.h"
+
+PRLogModuleInfo* GetLog()
+{
+  static PRLogModuleInfo* log = PR_NewLogModule("mozstumbler");
+  return log;
+}
